task0

+++ 0x01-variables_if_else_while/0-positive_or_negative.c
@@ -16,10 +16,10 @@ int main(void)
 	n = rand() - RAND_MAX / 2;
 	/* your code goes there */
 	if(n > 0)
-		printf("%u is positive\n", (unsigned int)n);
+		printf("%u is positive\n", n);
 	else if(n == 0)
-		printf("%u is zero\n", (unsigned int)n);
+		printf("%u is zero\n", n);
 	else
-		printf("%u is negative\n", (unsigned int)n);
+		printf("%u is negative\n", n);
 	return (0);
 }
